+++ src/main/java/zdoctor/morecrops/crops/Crops.java
@@ -1,9 +1,6 @@
 package zdoctor.morecrops.crops;
 
-import zdoctor.morecrops.crops.classes.DiamondCrop;
-import zdoctor.morecrops.crops.classes.EmeraldCrop;
-import zdoctor.morecrops.crops.classes.GoldCrop;
-import zdoctor.morecrops.crops.classes.IronCrop;
+import zdoctor.morecrops.crops.classes.*;
 
 public class Crops {
 	public static void load() {
@@ -11,5 +8,6 @@ public class Crops {
 		EmeraldCrop.load();
 		GoldCrop.load();
 		IronCrop.load();
+		StrawberryCrop.load();
 	}
 }