+++ src/main/java/zdoctor/morecrops/crops/classes/DiamondCrop.java
@@ -11,8 +11,8 @@ import net.minecraft.util.BlockPos;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.world.World;
 import net.minecraftforge.common.config.Configuration;
-import zdoctor.morecrops.GuiFactoryMoreCrops;
 import zdoctor.morecrops.MoreCrops;
+import zdoctor.morecrops.config.Config;
 import zdoctor.morecrops.gameregistry.ZItems;
 import zdoctor.zcore.common.EasyCrop;
 import zdoctor.zcore.common.EasySeed;
@@ -24,17 +24,23 @@ public class DiamondCrop {
 			@Override
 			public boolean canPlaceBlockAt(World worldIn, BlockPos pos) {
 				return super.canPlaceBlockAt(worldIn, pos)
-					&& HelperFunctions.checkSurrondingsFor(Blocks.lava, worldIn, true, pos);
+					&& HelperFunctions.checkSurrondingsFor(Blocks.lava, worldIn, pos, true);
 			};
 			@Override
 			public boolean canBlockStay(World worldIn, BlockPos pos, IBlockState state) {
-				return super.canBlockStay(worldIn, pos, state) && HelperFunctions.checkSurrondingsFor(Blocks.lava, worldIn, true, pos.down());
+				return super.canBlockStay(worldIn, pos, state) && HelperFunctions.checkSurrondingsFor(Blocks.lava, worldIn, pos.down(), true);
 			}
 			@Override
 			public boolean canUseBonemeal(World worldIn, Random rand, BlockPos pos, IBlockState state) {
-				return GuiFactoryMoreCrops.config.getBoolean("Allow Bone Meal Use", Configuration.CATEGORY_GENERAL, 
-						false, "Bone Meal Use");
+				return canUseBonemeal();
 			}
+			
+			@Override
+			public boolean canUseBonemeal() {
+				return !Config.disableBoneMeal.getBoolean(true);
+			}
+			
+			
 		};
 		DiamondPlant.growsOn(Blocks.coal_block);
 		
@@ -42,10 +48,15 @@ public class DiamondCrop {
 			@Override
 			public boolean onItemUse(ItemStack stack, EntityPlayer playerIn, World worldIn, BlockPos pos,
 					EnumFacing side, float hitX, float hitY, float hitZ) {
-				if(HelperFunctions.checkSurrondingsFor(Blocks.lava, worldIn, true, pos))
+				if(HelperFunctions.checkSurrondingsFor(Blocks.lava, worldIn, pos, true))
 					return super.onItemUse(stack, playerIn, worldIn, pos, side, hitX, hitY, hitZ);
 				return false;
 			};
+		
+			@Override
+			public boolean registerRecipe() {
+				return !Config.disableSeeds.getBoolean(false);
+			}
 		};
 		DiamondSeed.setRecipe(new Object[] {
 			"DDD", "DND", "DDD", 'D', Blocks.diamond_block, 'N', Items.nether_star