+++ src/main/java/zdoctor/morecrops/crops/classes/GoldCrop.java
@@ -7,9 +7,8 @@ import net.minecraft.init.Blocks;
 import net.minecraft.init.Items;
 import net.minecraft.util.BlockPos;
 import net.minecraft.world.World;
-import net.minecraftforge.common.config.Configuration;
-import zdoctor.morecrops.GuiFactoryMoreCrops;
 import zdoctor.morecrops.MoreCrops;
+import zdoctor.morecrops.config.Config;
 import zdoctor.morecrops.gameregistry.ZItems;
 import zdoctor.zcore.common.EasyCrop;
 import zdoctor.zcore.common.EasySeed;
@@ -19,12 +18,21 @@ public class GoldCrop {
 		EasyCrop GoldPlant = new EasyCrop("GoldCrop", MoreCrops.modid){
 			@Override
 			public boolean canUseBonemeal(World worldIn, Random rand, BlockPos pos, IBlockState state) {
-				return GuiFactoryMoreCrops.config.getBoolean("Allow Bone Meal Use", Configuration.CATEGORY_GENERAL, 
-						false, "Bone Meal Use");
+				return canUseBonemeal();
+			}
+			
+			@Override
+			public boolean canUseBonemeal() {
+				return !Config.disableBoneMeal.getBoolean(true);
 			}
 		};
 		
-		EasySeed GoldSeed = new EasySeed("GoldSeed", MoreCrops.modid);
+		EasySeed GoldSeed = new EasySeed("GoldSeed", MoreCrops.modid) {
+			@Override
+			public boolean registerRecipe() {
+				return !Config.disableSeeds.getBoolean(false);
+			}
+		};
 		GoldSeed.setRecipe(new Object[] {
 			"GGG", "GNG", "GGG", 'G', Blocks.gold_block, 'N', Items.nether_star
 		});